pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logawstudent"
version = "1.0.0"
description = "Automatiza el login y lanzamiento de laboratorios de AWS Academy."
authors = [
  { name = "Lazheart", email = "lazheart29@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "selenium",
  "python-dotenv",
  "webdriver-manager"
]

[project.scripts]
awstudent = "logawstudent.cli:app"
